Add tests for futures io as well
